==> src/01_data_gathering.ipynb <==
# Imports

import pandas as pd

from scraper import scrape
from downloader import download_and_extract_zips

DATA_PATH_ORIG = '../DATA/original'
DATA_PATH_PREP = '../DATA/prepared'

URL = 'https://chitanka.info/person'

AUTHORS = {
    'ivan_vazov': ["Българският език", "Отечество любезно, как хубаво си ти", "При Рилския манастир", "Елате ни вижте", "Линее нашто поколенье", "Левски", "Паисий", "Кочо", "Опълченците на Шипка", "Дядо Йоцо гледа", "Чичовци", "Под игото"],
    'aleko-konstantinov': ["Разни хора, разни идеали I", "Разни хора, разни идеали II", "Разни хора, разни идеали III", "Разни хора, разни идеали IV", "Бай Ганьо"],
    'jordan-jovkov': ["Песента на колелетата", "Последна радост", "Шибил", "През чумавото", "Индже", "Албена", "Другоселец", "Серафим"],
    'elin-pelin': ["Ветрената мелница", "Косачи", "Задушница", "Мечтатели", "На оня свят", "Андрешко", "Чорба от греховете на отец Никодим", "Занемелите камбани", "Гераците"],
    'dimityr-dimov': ["Тютюн"],
    'dimityr-talev': ["Железният светилник"],
}

# Get scraped data

df = scrape(URL, AUTHORS.keys())
df

all_texts = sum(AUTHORS.values(), [])
all_texts[:5]

# check whether all texts you're going to use are present in the dataframe
all(all(text in set(df.query('author == author')['title']) for text in AUTHORS[author]) for author in AUTHORS)

# Subset only texts of interest

df_full = df.query('title in @all_texts').reset_index(drop=True)
df_full

for auth in AUTHORS:
    print(auth, end=' => ')
    print(set(df_full.query("author == @auth")['title']) == set(AUTHORS[auth]), end=' | ')
    print(len(df_full.query("author == @auth")['title']) == len(AUTHORS[auth]), end='')
    print()

# jordan-jovkov has Албена as an "Пиеса" and as a "Разказ". We need the second one.
# Its download link includes 7879. Therefore, drop row with index 18.
# There are also two vesions of Серафим. Not sure why. Keeping one of them. Drop row with index 25.

# elin-pelin has Косачи as an "Поезия" and as a "Разказ". We need the second one.
# Its download link includes 7879. Therefore, drop row with index 28.

# dimityr-dimov has two versions of "Тютюн" due to political reasons. We'll work with the most recent one.
# Its download link includes 2760. Therefore, drop row with index 38.

df_full = df_full.drop([18, 25, 28, 38])

for auth in AUTHORS:
    print(auth, end=' => ')
    print(set(df_full.query("author == @auth")['title']) == set(AUTHORS[auth]), end=' | ')
    print(len(df_full.query("author == @auth")['title']) == len(AUTHORS[auth]), end='')
    print()

df_full

# Saving to files

df_full.to_pickle(f'{DATA_PATH_PREP}/01_df_full_links.pkl')
