==> src/02_data_loading.ipynb <==
# Imports

%load_ext autoreload
%autoreload 2

from typing import List

import re
import requests

import pandas as pd

from scraper import scrape
from downloader import download_and_extract_zips

DATA_PATH_ORIG = '../DATA/original'
DATA_PATH_PREP = '../DATA/prepared'

URL = 'https://chitanka.info/person'

AUTHORS = {
    'ivan_vazov': ["Българският език", "Отечество любезно, как хубаво си ти", "При Рилския манастир", "Елате ни вижте", "Линее нашто поколенье", "Левски", "Паисий", "Кочо", "Опълченците на Шипка", "Дядо Йоцо гледа", "Чичовци", "Под игото"],
    'aleko-konstantinov': ["Разни хора, разни идеали I", "Разни хора, разни идеали II", "Разни хора, разни идеали III", "Разни хора, разни идеали IV", "Бай Ганьо"],
    'jordan-jovkov': ["Песента на колелетата", "Последна радост", "Шибил", "През чумавото", "Индже", "Албена", "Другоселец", "Серафим"],
    'elin-pelin': ["Ветрената мелница", "Косачи", "Задушница", "Мечтатели", "На оня свят", "Андрешко", "Чорба от греховете на отец Никодим", "Занемелите камбани", "Гераците"],
    'dimityr-dimov': ["Тютюн"],
    'dimityr-talev': ["Железният светилник"],
}

# Load data

df_full = pd.read_pickle(f'{DATA_PATH_PREP}/01_df_full_links.pkl')
df_full

# Download data

urls = df_full['download_link'].tolist()
names = download_and_extract_zips(urls, DATA_PATH_ORIG)

df_full['local_filename'] = names
df_full

# Add column with the text for each title

def get_text(filename: str) -> str:
    with open(filename, 'r', encoding='utf-8') as f:
        content = f.read()
    return content

df_full['text'] = df_full['local_filename'].map(get_text)
df_full.head()

# Remove metainformation

- At the beginning of each text there is name of the author and the name of the text. We already have this data in the dataframe, so we can ignore it now.
- At the end of each text there is metainformation about the text: who digitalized it, from where it was taken, etc. Remove it.

def remove_suffix(content: str) -> str:
    idx_end = content.rfind('$id')
    return content[:idx_end]

def remove_prefix(content: str) -> str:
    idx_newline_fst = content.find("\n")
    idx_newline_snd = content.find("\n", idx_newline_fst + 1)
    return content[idx_newline_snd:]

## Suffix

tmp = df_full['text'][0]
idx_end = len(tmp) - tmp.rfind('$id')
print(f'{idx_end=}')

print('################ Before ################')
print(df_full['text'][0][-idx_end-20:])
print('########################################')

df_full['text'] = df_full['text'].map(remove_suffix)
df_full.head()

print('################ After ################')
print(df_full['text'][0][-idx_end:])
print('########################################')

## Prefix

print('################ Before ################')
print(df_full['text'][0][:100])
print('########################################')

df_full['text'] = df_full['text'].map(remove_prefix)
df_full.head()

print('################ After ################')
print(df_full['text'][0][:100])
print('########################################')

df_full.head()

# Saving to files

df_full.to_pickle(f'{DATA_PATH_PREP}/02_df_full_text_no_meta.pkl')
