==> src/04_embeddings.ipynb <==
# Imports

import pickle

import pandas as pd
pd.set_option('max_colwidth', 150)
from sklearn.feature_extraction.text import TfidfVectorizer

DATA_PATH_PREP = '../DATA/prepared'

# Load data

df_full_soft = pd.read_pickle(f'{DATA_PATH_PREP}/03_df_full_soft.pkl')
df_full_med = pd.read_pickle(f'{DATA_PATH_PREP}/03_df_full_med.pkl')
df_full_hard = pd.read_pickle(f'{DATA_PATH_PREP}/03_df_full_hard.pkl')

df_full_soft.head()

# Term Frequency - Inverse Document Frequency Table

def identity(x):
    return x


tfidf_soft = TfidfVectorizer(
    tokenizer=identity,
    preprocessor=identity,
    token_pattern=None,
    lowercase=False,
    stop_words=None,
    min_df=2,
)

data_tfidf = tfidf_soft.fit_transform(df_full_soft['text'])
data_table_soft = pd.DataFrame(
    data_tfidf.toarray(),
    columns=tfidf_soft.get_feature_names_out(),
    index=df_full_soft.index
)
data_table_soft['author'] = df_full_soft['author']
data_table_soft.head()

tfidf_med = TfidfVectorizer(
    tokenizer=identity,
    preprocessor=identity,
    token_pattern=None,
    lowercase=False,
    stop_words=None,
    min_df=2
)

data_tfidf = tfidf_med.fit_transform(df_full_med['text'])
data_table_med = pd.DataFrame(
    data_tfidf.toarray(),
    columns=tfidf_med.get_feature_names_out(),
    index=df_full_med.index
)
data_table_med['author'] = df_full_med['author']
data_table_med.head()

tfidf_hard = TfidfVectorizer(
    tokenizer=identity,
    preprocessor=identity,
    token_pattern=None,
    lowercase=False,
    stop_words=None,
    min_df=2,
)

data_tfidf = tfidf_hard.fit_transform(df_full_hard['text'])
data_table_hard = pd.DataFrame(
    data_tfidf.toarray(),
    columns=tfidf_hard.get_feature_names_out(),
    index=df_full_hard.index
)
data_table_hard['author'] = df_full_hard['author']
data_table_hard.head()

# Saving to files

data_table_soft.to_pickle(f'{DATA_PATH_PREP}/04_data_table_soft.pkl')
data_table_med.to_pickle(f'{DATA_PATH_PREP}/04_data_table_med.pkl')
data_table_hard.to_pickle(f'{DATA_PATH_PREP}/04_data_table_hard.pkl')

pickle.dump(tfidf_soft, open(f'{DATA_PATH_PREP}/04_vectorizer_soft.pkl', 'wb'))
pickle.dump(tfidf_med, open(f'{DATA_PATH_PREP}/04_vectorizer_med.pkl', 'wb'))
pickle.dump(tfidf_hard, open(f'{DATA_PATH_PREP}/04_vectorizer_hard.pkl', 'wb'))
