==> src/06_modelling.ipynb <==
# Imports

import numpy as np
import pandas as pd
pd.set_option('max_colwidth', 150)

import pickle

from sklearn.model_selection import train_test_split, StratifiedKFold, cross_val_score
from sklearn.preprocessing import LabelEncoder

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import SVC
from catboost import CatBoostClassifier


from sklearn.metrics import accuracy_score, log_loss, f1_score, matthews_corrcoef

DATA_PATH_PREP = '../DATA/prepared'

# Load data

# data_table_soft = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_soft.pkl')
# data_table_med = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_med.pkl')
data_table_hard = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_hard.pkl')
data_table_hard

# Split data

X = data_table_hard.drop('author', axis=1)
y = data_table_hard['author']

le = LabelEncoder()
y = le.fit_transform(y)
print(dict(enumerate(le.classes_)))

seed = 42
X_train, X_test_val_test, y_train, y_test_val_test = train_test_split(X, y, random_state=seed, train_size=0.8)
X_val, X_test, y_val, y_test = train_test_split(X_test_val_test, y_test_val_test, random_state=seed, train_size=0.5)

print(f'{X_train.shape} | {y_train.shape}')
print(f'{X_val.shape}  | {y_val.shape}')
print(f'{X_test.shape}  | {y_test.shape}')

# Choose a model

strat_k_fold = StratifiedKFold(n_splits=10, shuffle=True)

lr = LogisticRegression()
print(f'LogisticRegression: {cross_val_score(lr, X, y, cv=strat_k_fold).mean()}')

knn = KNeighborsClassifier(n_neighbors=5)
print(f'KNeighborsClassifier: {cross_val_score(knn, X, y, cv=strat_k_fold).mean()}')

gnb = GaussianNB()
print(f'GaussianNB: {cross_val_score(gnb, X, y, cv=strat_k_fold).mean()}')

mnb = MultinomialNB()
print(f'MultinomialNB: {cross_val_score(mnb, X, y, cv=strat_k_fold).mean()}')

svc = SVC()
print(f'SVC: {cross_val_score(svc, X, y, cv=strat_k_fold).mean()}')

## Explore `CatBoost`

cat_boost = CatBoostClassifier()
cat_boost.fit(X_train, y_train)

y_pred = cat_boost.predict(X_val)
y_pred_proba = cat_boost.predict_proba(X_val)

acc = accuracy_score(y_val, y_pred)
neg_log_loss = log_loss(y_val, y_pred_proba)
mcc = matthews_corrcoef(y_val, y_pred)
f1 = f1_score(y_val, y_pred, average='macro')

print(f'{acc=}' )
print(f'{neg_log_loss=}' )
print(f'{mcc=}' )
print(f'{f1=}' )

## Explore `LogisticRegression`

log_reg = LogisticRegression()
log_reg.fit(X_train, y_train)

y_pred = log_reg.predict(X_val)
y_pred_proba = log_reg.predict_proba(X_val)

acc = accuracy_score(y_val, y_pred)
neg_log_loss = log_loss(y_val, y_pred_proba)
mcc = matthews_corrcoef(y_val, y_pred)
f1 = f1_score(y_val, y_pred, average='macro')

print(f'{acc=}' )
print(f'{neg_log_loss=}' )
print(f'{mcc=}' )
print(f'{f1=}' )

# Evaluate

y_pred = cat_boost.predict(X_test)
y_pred_proba = cat_boost.predict_proba(X_test)

acc = accuracy_score(y_test, y_pred)
neg_log_loss = log_loss(y_test, y_pred_proba)
mcc = matthews_corrcoef(y_test, y_pred)
f1 = f1_score(y_test, y_pred, average='macro')

print(f'{acc=}' )
print(f'{neg_log_loss=}' )
print(f'{mcc=}' )
print(f'{f1=}' )

y_pred = log_reg.predict(X_test)
y_pred_proba = log_reg.predict_proba(X_test)

acc = accuracy_score(y_test, y_pred)
neg_log_loss = log_loss(y_test, y_pred_proba)
mcc = matthews_corrcoef(y_test, y_pred)
f1 = f1_score(y_test, y_pred, average='macro')

print(f'{acc=}' )
print(f'{neg_log_loss=}' )
print(f'{mcc=}' )
print(f'{f1=}' )

# Saving to files

pickle.dump(cat_boost, open(f'{DATA_PATH_PREP}/06_model_cat_boost.pkl', 'wb'))
pickle.dump(log_reg, open(f'{DATA_PATH_PREP}/06_model_log_reg.pkl', 'wb'))
