==> src/06_modelling03_pycaret.ipynb <==
# Imports

import numpy as np
import pandas as pd
pd.set_option('max_colwidth', 150)
import matplotlib.pyplot as plt

from pycaret.classification import *

from catboost import CatBoostClassifier

from sklearn.metrics import accuracy_score, log_loss, f1_score, matthews_corrcoef

DATA_PATH_PREP = '../DATA/prepared'

# Load data

# data_table_soft = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_soft.pkl')
# data_table_med = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_med.pkl')
data_table_hard = pd.read_pickle(f'{DATA_PATH_PREP}/04_data_table_hard.pkl')
data_table_hard

# Choose a model with PyCaret

## With PCA

s = setup(
    data=data_table_hard,
    target='author',
    train_size=0.8,
    pca=True,
    pca_method='linear',
    # remove_multicollinearity=True,
)

remove_metric('AUC')
remove_metric('Recall')
remove_metric('Precision')
add_metric('logloss', 'LogLoss', log_loss, greater_is_better=False, target='pred_proba')

best = compare_models(sort='LogLoss')

## Without PCA

s = setup(
    data=data_table_hard,
    target='author',
    train_size=0.8,
    # pca=True,
    # pca_method='linear',
    # remove_multicollinearity=True,
)

remove_metric('AUC')
remove_metric('Recall')
remove_metric('Precision')
add_metric('logloss', 'LogLoss', log_loss, greater_is_better=False, target='pred_proba')

best = compare_models(sort='LogLoss')

# Choosing a model

s = setup(
    data=data_table_hard,
    target='author',
    train_size=0.8,
    pca=True,
    pca_method='linear',
    # remove_multicollinearity=True,
)

remove_metric('AUC')
remove_metric('Recall')
remove_metric('Precision')
add_metric('logloss', 'LogLoss', log_loss, greater_is_better=False, target='pred_proba')

ridge = create_model('ridge', return_train_score=True)
final_ridge = s.finalize_model(ridge)
final_ridge

# Saving to files

filename = f'{DATA_PATH_PREP}/06_pycaret_ridge'
save_model(final_ridge, filename)
